% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_twind.R
\name{calc_twind}
\alias{calc_twind}
\title{Calculate total wind speed from zonal and meridional components}
\usage{
calc_twind(uwind, vwind)
}
\arguments{
\item{uwind}{numeric; the zonal wind component}

\item{vwind}{numeric; the meridional wind component}
}
\description{
Obtain total wind speed using the zonal (u) and meridonal (v) wind components.
}
