% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_winddir.R
\name{calc_winddir}
\alias{calc_winddir}
\title{Calculate wind direction from zonal and meridional components}
\usage{
calc_winddir(uwind, vwind)
}
\arguments{
\item{uwind}{numeric; the zonal wind component}

\item{vwind}{numeric; the meridional wind component}
}
\description{
Obtain wind direction in degrees using the zonal (u) and meridonal (v) wind components.
See https://confluence.ecmwf.int/pages/viewpage.action?pageId=133262398 or equations in e.g., https://doi.org/10.3390/geosciences9070289
}
