% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_era5.R
\name{collate_era5}
\alias{collate_era5}
\title{Collate a set of ERA5 files given as a data frame of filenames.}
\usage{
collate_era5(filenames, string = NULL, coords = NULL)
}
\arguments{
\item{filenames}{A data frame containing a single column of filenames.}

\item{string}{NULL or a single character string used to identify target file types within the filenames.}

\item{coords}{NULL or a character string containing an individual lat, lon grid position e.g. c(-80.25, -81.25)}
}
\description{
A fetcher/parser for single netcdf4-formatted ERA5 downloaded from the CDS.
}
