% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_era5.R
\name{extract_era5}
\alias{extract_era5}
\title{Fetch a single ERA5 file. Option to extract specific coordinates.}
\usage{
extract_era5(nc_filepath, coords = NULL)
}
\arguments{
\item{nc_filepath}{full file path to the target netcdf4 era5 file.}

\item{coords}{a character string containing an individual lat, lon grid position e.g. c(-80.25, -81.25)}
}
\description{
A fetcher/parser for single netcdf4-formatted ERA5 downloaded from the CDS.
}
