% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CDS_era5.R
\name{get_CDS_era5}
\alias{get_CDS_era5}
\title{Fetch ERA-5 nc files containing variables needed for an accumulation}
\usage{
get_CDS_era5(
  key,
  user,
  coords = c(-79.25, -161.75, -79.5, -161.5),
  by = "month",
  variables = c("2m_temperature", "total_precipitation"),
  start_YYYYMM = "197901",
  end_YYYYMM = "202412",
  identifier = "SITE",
  download_directory,
  archive = TRUE
)
}
\arguments{
\item{key}{Copernicus CDS API key. Register with the ECMWF then go to https://cds.climate.copernicus.eu/profile.}

\item{user}{email address associated with the above API key.}

\item{coords}{a vector of latitude and longitude for a bounding box to use when fetching. Refer to the web request form. Order is c(N,W,S,E).}

\item{by}{individual request size. Only "month" supported currently, as it is recommended for being the most efficient.}

\item{variables}{strings denoting variables to get. Refer to \code{\link[ecmwfr:wf_request]{ecmwfr::wf_request()}}.}

\item{start_YYYYMM}{A string for starting year-month in the format "YYYYMM"}

\item{end_YYYYMM}{A string for ending year-month in the format "YYYYMM"}

\item{identifier}{A string to identify the files. E.g., project/site name, variables, etc. Files will always start with "ERA5-".}

\item{download_directory}{a filepath to the folder where files will be downloaded to.}

\item{archive}{TRUE/FALSE to archive existing files within the supplied download_directory}
}
\description{
Use ecmwfr to interface with the Copernicus API and obtain ERA-5 netcdf4 files for use by \code{accumulate()}
}
