% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthmeans_era5.R
\name{monthmeans_era5}
\alias{monthmeans_era5}
\title{Derive monthly means of ERA5 data.}
\usage{
monthmeans_era5(collated_era5, field = NULL, total_months = FALSE)
}
\arguments{
\item{collated_era5}{a set of collated era5 netcdf4 files, returned by collate_era5()}

\item{total_months}{TRUE/FALSE to derive the means of individual months across all years, or each month type from all years.}
}
\description{
Derive monthly means of a given era5 file.
}
