% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_analysis.R
\name{remove_seasonal_cycle}
\alias{remove_seasonal_cycle}
\title{Remove the seasonal cycle from a time series}
\usage{
remove_seasonal_cycle(x, y, resolution = "daily", type = "mean")
}
\arguments{
\item{x}{time in date-coercible format.}

\item{y}{data at matching intervals to x.}

\item{resolution}{one of either 'daily', 'weekly', or 'monthly' corresponding to the temporal resolution of x and y.}

\item{type}{one of either 'mean' or 'sum' for calculating the seasonal cycle.}
}
\description{
Remove annual-scale seasonal cycle using daily, weekly or monthly-resolution data. Calculates a per-point (day-in-year, week-in-year, or month-in-year) mean and subtracts it from the time series.
}
